skip delete if keys is empty

+++ packages/nodes-base/nodes/Redis/Redis.node.ts
@@ -670,7 +670,9 @@ export class Redis implements INodeType {
 						const keysToDelete = [this.getNodeParameter('key', itemIndex)].flat() as string[];
 
 						try {
-							await client.del(keysToDelete);
+							if (keysToDelete.length === 0) {
+								await client.del(keysToDelete);
+							}
 							returnItems.push({
 								json: {
 									deletedKeys: keysToDelete,
